simplify sorting example
