+++ src/MUIDataTable.js
@@ -703,7 +703,7 @@ class MUIDataTable extends React.Component {
   }
 
   sortTable(data, col, order) {
-    let dataSrc = this.options.customSort ? this.options.customSort(data, col, order) : data;
+    let dataSrc = this.options.customSort ? this.options.customSort(data, col, order || "asc") : data;
 
     let sortedData = dataSrc.map((row, sIndex) => ({
       data: row.data[col],
